Switch among two accumulates
